feat(clients): model and view clean up
