+++ app/tests/spec/models/attached-clients.js
